node.cpp
Code that runs on each of the stationary zigbee nodes. It listens to the boebot and responds with its ID as acknowledgment only when the Boebot sends a 'P' (Ping signal) along with the nodeID it is searching for. It stops sending acknowledgment when the Boebot receives the acknowledgment and sense out 'S' (Stop signal).